refactor(sslib): fold sslib/ back into utils.py and delete sub-package (#177)
Move all 4 sslib modules (config, aws_client, concurrency, cost) directly
into utils.py under labeled section headers. Removes the Sprint 4 compatibility
shim block and eliminates the sslib package entirely.

Update tests to import from utils instead of sslib.*; rewrote patch targets
throughout test_utils.py, test_duplicate_code_refactoring.py, and tests/sslib/.
