+++ pyproject.toml
@@ -89,7 +89,7 @@ Issues = "https://github.com/ColonelPanicX/StratusScan-CLI/issues"
 
 # Setuptools configuration
 [tool.setuptools]
-packages = ["sslib", "scripts.smart_scan"]
+packages = ["scripts.smart_scan"]
 include-package-data = true
 
 [tool.setuptools.package-data]
@@ -173,11 +173,6 @@ module = [
     "openpyxl.*",
 ]
 ignore_missing_imports = true
-
-# Enforce strict typing on the sslib/ package (Sprint 4+)
-[[tool.mypy.overrides]]
-module = ["sslib.*"]
-disallow_untyped_defs = true
 
 # Pytest configuration
 [tool.pytest.ini_options]
