--- sslib/__init__.py
@@ -1,1 +1,0 @@
-# sslib — StratusScan shared library package
